requirements.txt
astroid>=2.2.5
Click>=7.0
Flask>=1.0.3
isort>=4.3.20
itsdangerous>=1.1.0
Jinja2>=2.10.1
lazy-object-proxy>=1.4.1
MarkupSafe>=1.1.1
mccabe>=0.6.1
mysql-connector-python>=8.0.18
numpy>=1.16.4
protobuf>=3.9.1
python-dateutil>=2.8.0
python-slugify>=3.0.6
pytz>=2019.3
six>=1.12.0
text-unidecode>=1.3
typed-ast>=1.4.0
Werkzeug>=0.15.4
wrapt>=1.11.1
uWSGI>=2.0.18
